requirements/dev.txt
# python version = 3.10.11
# python -m pip install --upgrade pip

rich
sqlalchemy
python-dotenv

wtforms
email_validator
werkzeug

bcrypt
cryptography
pyjwt[crypto]


